requirements.txt
requests
beautifulsoup4
schedule
pysocks
tabulate
reportlab
python-telegram-bot
flask
